refactor: group navigation and traffic under features
