+++ src/features/navigation/services/routeCache.ts
