+++ src/features/navigation/ui/__tests__/DrivingHUD.test.tsx
@@ -8,7 +8,7 @@ jest.mock('react-native', () => ({
   Text: 'Text',
   StyleSheet: { create: () => ({}), flatten: () => ({}) },
 }));
-jest.mock('../../premium/subscription', () => ({
+jest.mock('../../../../premium/subscription', () => ({
   usePremium: () => ({ isPremium: true }),
 }));
