+++ src/features/navigation/ui/__tests__/SpeedBanner.test.tsx
