+++ src/features/traffic/__tests__/trafficLightDetector.test.ts
