+++ src/features/traffic/trafficLightDetector.ts
